+++ make_responsive_images/__init__.py
@@ -1,2 +1,2 @@
 __app_name__ = "make-responsive-images"
-__version__ = "0.1.7"
+__version__ = "0.1.8"